requirements.txt
customtkinter>=5.0.0
Pillow>=9.0.0
cryptography>=41.0.0
psutil>=5.9.0
bcrypt>=4.0.0
pyinstaller>=6.0.0
